2/13/23
2/13/23 - Changed Voltage, changing the usb to http in Robot, got rid of buttons in RobotContainer since we aren't going to use them. Commented out all getters for the motors in ShuffleboardUpdater since the bot doesn't take them. Updated LimelightHelpers and got rid of the rest.
