+++ src/main/java/frc/robot/Constants.java
@@ -24,10 +24,10 @@ public final class Constants {
     public static int frontLeftDrive = 4; //6
     public static int frontRightDrive = 7; //7
     public static int backRightDrive = 8; //8
-    public static int clawMotorLeft = 7;
-    public static int clawMotorRight = 8;
+    public static int clawMotorLeft = 6;
+    public static int clawMotorRight = 9;
     public static int towerMotor = 1;
-    public static int armMotor = 6;
+    public static int armMotor = 3;
 
     //the number is the CAN ID that the spark max is attached to :)
     //public static int towerExtension = 7;
@@ -46,8 +46,8 @@ public final class Constants {
     //change this later when figuring out the voltage for the robot
     //claw volt
 
-    public static double clawMotorLeftVolt = 7;
-    public static double clawMotorRightVolt = 7;
+    public static double clawMotorLeftVolt = 3;
+    public static double clawMotorRightVolt = 3;
     //drive volt
     public static double autoDrive = .3;
     //tower volt
