--- src/main/java/frc/robot/utils/LimelightResults.java
@@ -1,19 +1,0 @@
-package frc.robot.utils;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.Setter;
-
-public class LimelightResults {
-
-    @Getter
-    @Setter
-    @JsonProperty("Results")
-    private Results results;
-
-    public LimelightResults() {
-        results = new Results();
-    }
-
-
-}
