--- src/main/java/frc/robot/utils/LimelightTargetBarcode.java
@@ -1,5 +1,0 @@
-package frc.robot.utils;
-
-class LimelightTargetBarcode {
-
-}
